+++ confluent_kafka-stubs/admin/__init__.pyi
@@ -2,95 +2,3 @@
 types-confluent-kafka: A package providing type hints for the confluent-kafka Python package.
 This package is licensed under the Apache 2.0 License.
 """
-
-# standard library
-from enum import Enum
-from typing import TYPE_CHECKING, Any, Literal
-
-from .._util import ConversionUtil, ValidationUtil
-from ._resource import ResourcePatternType, ResourceType
-
-if TYPE_CHECKING: ...
-
-try:
-    string_type = basestring  # pyright: ignore
-except NameError:
-    string_type = str
-
-class AclOperation(Enum):
-    UNKNOWN: Literal["ACL_OPERATION_UNKNOWN"]
-    ANY: Literal["ACL_OPERATION_ANY"]
-    ALL: Literal["ACL_OPERATION_ALL"]
-    READ: Literal["ACL_OPERATION_READ"]
-    WRITE: Literal["ACL_OPERATION_WRITE"]
-    CREATE: Literal["ACL_OPERATION_CREATE"]
-    DELETE: Literal["ACL_OPERATION_DELETE"]
-    ALTER: Literal["ACL_OPERATION_ALTER"]
-    DESCRIBE: Literal["ACL_OPERATION_DESCRIBE"]
-    CLUSTER_ACTION: Literal["ACL_OPERATION_CLUSTER_ACTION"]
-    DESCRIBE_CONFIGS: Literal["ACL_OPERATION_DESCRIBE_CONFIGS"]
-    ALTER_CONFIGS: Literal["ACL_OPERATION_ALTER_CONFIGS"]
-    IDEMPOTENT_WRITE: Literal["ACL_OPERATION_IDEMPOTENT_WRITE"]
-    def __lt__(self, other) -> bool: ...
-
-class AclPermissionType(Enum):
-    UNKNOWN: Literal["ACL_PERMISSION_TYPE_UNKNOWN"]
-    ANY: Literal["ACL_PERMISSION_TYPE_ANY"]
-    DENY: Literal["ACL_PERMISSION_TYPE_DENY"]
-    ALLOW: Literal["ACL_PERMISSION_TYPE_ALLOW"]
-    def __lt__(self, other) -> bool: ...
-
-class AclBinding:
-    def __init__(
-        self,
-        restype: "ResourceType",
-        name: str,
-        resource_pattern_type: "ResourcePatternType",
-        principal: str,
-        host: str,
-        operation: "AclOperation",
-        permission_type: "AclPermissionType",
-    ) -> None: ...
-    def _convert_enums(self): ...
-    def _check_forbidden_enums(self, forbidden_enums): ...
-    def _not_none_args(
-        self,
-    ) -> Literal["restype", "name", "resource_pattern_type", "principal", "host", "operation", "permission_type"]: ...
-    def _string_args(self) -> Literal["name", "principal", "host"]: ...
-    def _forbidden_enums(
-        self,
-    ) -> dict[
-        str,
-        Literal[
-            "RESOURCE_ANY",
-            "RESOURCE_PATTERN_ANY",
-            "RESOURCE_PATTERN_MATCH",
-            "ACL_OPERATION_ANY",
-            "ACL_PERMISSION_TYPE_ANY",
-        ],
-    ]: ...
-    def _convert_args(self) -> None: ...
-    def _to_tuple(self) -> tuple[Any]: ...
-    def __repr__(self) -> str: ...
-    def __hash__(self) -> int: ...
-    def __lt__(self, other) -> bool: ...
-    def __eq__(self, other) -> bool: ...
-
-class AclBindingFilter(AclBinding):
-    def __init__(
-        self,
-        restype: "ResourceType",
-        name: str,
-        resource_pattern_type: "ResourcePatternType",
-        principal: str,
-        host: str,
-        operation: "AclOperation",
-        permission_type: "AclPermissionType",
-    ) -> None: ...
-    def _not_none_args(self) -> Literal["restype", "resource_pattern_type", "operation", "permission_type"]: ...
-    def _forbidden_enums(
-        self,
-    ) -> dict[
-        str,
-        Literal["RESOURCE_UNKNOWN", "RESOURCE_PATTERN_UNKNOWN", "ACL_OPERATION_UNKNOWN", "ACL_PERMISSION_TYPE_UNKNOWN"],
-    ]: ...
